new_blog/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.3
Cython==0.29.14
distlib==0.3.2
Django==3.0.6
django-crispy-forms==1.12.0
django-froala-editor==3.2.2
django-multiselectfield==0.1.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
filelock==3.0.12
gensim==3.8.3
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.3
Pillow==7.1.2
PyJWT==2.1.0
pylint==2.5.2
python-dotenv==0.18.0
pytz==2020.1
requests==2.25.1
scipy==1.7.0
six==1.15.0
smart-open==5.1.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.6
virtualenv==20.4.7
wrapt==1.12.1
